pyproject.toml
[project]
name = "dumb-pass"
version = "0.1.0"
description = "Backend for strong-pass"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorlog>=6.9.0",
    "cryptography>=46.0.2",
    "fastapi[standard]",
    "psycopg2-binary>=2.9.10",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "ruff>=0.14.0",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
